tests/submission/PhilipTaylor/canvas/2d.state.saverestore.strokeStyle.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.state.saverestore.strokeStyle</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.state.html">state</a>.<a href="index.2d.state.saverestore.html">saverestore</a>.strokeStyle</h1>
<p class="desc">save()/restore() works for strokeStyle</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.state.strokeStyle">2d.state.strokeStyle</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("save()/restore() works for strokeStyle");
_addTest(function(canvas, ctx) {

// Test that restore() undoes any modifications
var old = ctx.strokeStyle;
ctx.save();
ctx.strokeStyle = "#ff0000";
ctx.restore();
_assertSame(ctx.strokeStyle, old, "ctx.strokeStyle", "old");

// Also test that save() doesn't modify the values
ctx.strokeStyle = "#ff0000";
old = ctx.strokeStyle;
    // we're not interested in failures caused by get(set(x)) != x (e.g.
    // from rounding), so compare against 'old' instead of against "#ff0000"
ctx.save();
_assertSame(ctx.strokeStyle, old, "ctx.strokeStyle", "old");
ctx.restore();


});
</script>

<div id="log"></div>
</body></html>
